Move the image encoder to detectron2 backbone and fix for tf.keras backend
