+++ keras_cv/models/segmentation/__init__.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from keras_cv.models.segmentation.deeplab_v3_plus import DeepLabV3Plus
-from keras_cv.models.segmentation.segment_anything import ImageEncoder
 from keras_cv.models.segmentation.segment_anything import MaskDecoder
 from keras_cv.models.segmentation.segment_anything import PromptEncoder
 from keras_cv.models.segmentation.segment_anything import TwoWayTransformer
