+++ keras_cv/models/segmentation/segment_anything/__init__.py
@@ -12,9 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_cv.models.segmentation.segment_anything.sam_image_encoder import (
-    ImageEncoder,
-)
 from keras_cv.models.segmentation.segment_anything.sam_mask_decoder import (
     MaskDecoder,
 )
