+++ keras_cv/models/segmentation/segment_anything/sam_layers.py
@@ -37,9 +37,6 @@ class MLPBlock(keras.layers.Layer):
         self.mlp_dim = mlp_dim
         self.activation = activation
 
-        self.built = False
-
-    def build(self, input_shape=None):
         self.dense_layer1.build([self.embedding_dim])
         self.dense_layer2.build([self.mlp_dim])
